+++ wazimap/models/__init__.py
@@ -1,0 +1,2 @@
+from .geo import GeographyBase, GeoMixin, Geography  # noqa
+from .data import FieldTable, SimpleTable, DBTable, Dataset, Release, FieldTableRelease, SimpleTableRelease  # noqa
